move UnionFind
